# openapi.yaml
openapi: 3.0.1
info:
  title: E-Commerce API
  version: 1.0.0
  description: A Basic CRUD API Example for an online store
paths:
...
  /api/cart:
    summary: Api Route for User's Carts
    get:
      summary: Get the data of all user's carts
      description: ''
      operationId: Get_carts
      responses:
        '200':
          description: An Object of Carts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The carts ID.
                  user_id:
                    type: integer
                    description: The user's id.
                  created_at:
                    type: string
                    description: A timestamp of the cart's creation.
                  
        default:
          description: Default error sample response
      tags:
        - Cart
    post:
      summary: Create a New Cart
      description: User Cart generation
      operationId: Create_Cart
      requestBody:
          description: The Cart must be attached to a user via their user id.
          content:
            application/json: 
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    description: The user ID.
      responses:
      
        default:
          description: Default error sample response
      tags:
        - Cart
  /api/cart/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: {}
    summary: CRUD operations for a specified cart
    get:
      summary: Get specified cart data
      description: ''
      operationId: Get_specific_cart
      responses:
        '200':
          description: An Object of Carts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The carts ID.
                  user_id:
                    type: integer
                    description: The user's id.
                  created_at:
                    type: string
                    description: A timestamp of the cart's creation.
        default:
          description: Default error sample response
      tags:
        - Cart
    delete:
      summary: Delete an existing cart record
      description: ''
      operationId: delete_cart
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Success:
                  value: Cart Record Successfully Deleted
        default:
          description: Default error sample response
      tags:
        - Cart
...
  /api/cart_item:
    summary: The products in each individual cart
    get:
      summary: Get a list of every item in every cart
      description: ''
      operationId: get_cart_item_list
      responses:
        '200':
          description: A User object
          content:
            application/json:
              schema:
                type: object
                properties:
                  product_id:
                    type: integer
                    description: The user ID.
                  cart_id:
                    type: integer
                    description: The user's first name.
                  quantity:
                    type: integer
                    description: The user's last name.
        default:
          description: Default error sample response
      tags:
        - Cart_Item
  /api/cart_item/{product_id}:
    parameters:
      - in: path
        name: product_id
        required: true
        schema: {}
    post:
      summary: Add an item to a cart
      description: ''
      operationId: add_cart_item
      requestBody:
          description: Adding an item to a cart
          content:
            application/json: 
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The order ID.
                  user_id:
                    type: integer
                    description: Id representing the owner of this order.
                  cart_id:
                    type: integer
                    description: Id representing the owner's cart which makes up the order.
                  price_total:
                    type: integer
                    description: Amount of all the items in the cart.
                  complete:
                    type: boolean
                    description: If the order has been completed or not.
                  completed_at:
                    type: integer
                    description: Timestamp of when the order was completed if it was.
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Success:
                  value: Item Successfully Added to Cart
        default:
          description: Default error sample response
      tags:
        - Cart_Item
    put:
      summary: Update Quantity of item in cart
      description: ''
      operationId: Update_Quantity
      requestBody:
          description: Updating an item in a user's cart
          content:
            application/json: 
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    description: Id representing the owner of this order.
                  cart_id:
                    type: integer
                    description: Id representing the owner's cart which makes up the order.
                  price_total:
                    type: integer
                    description: Amount of all the items in the cart.
                  complete:
                    type: boolean
                    description: If the order has been completed or not.
                  completed_at:
                    type: integer
                    description: Timestamp of when the order was completed if it was.
      responses:
        '200':
          description: A User object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The order ID.
                  user_id:
                    type: integer
                    description: Id representing the owner of this order.
                  cart_id:
                    type: integer
                    description: Id representing the owner's cart which makes up the order.
                  price_total:
                    type: integer
                    description: Amount of all the items in the cart.
                  complete:
                    type: boolean
                    description: If the order has been completed or not.
                  completed_at:
                    type: integer
                    description: Timestamp of when the order was completed if it was.
        default:
          description: Default error sample response
      tags:
        - Cart_Item
    delete:
      summary: Delete an item from cart
      description: ''
      operationId: delete_Cart_Item
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Success:
                  value: Item Successfully Deleted from Cart
        default:
          description: Default error sample response
      tags:
        - Cart_Item
  /api/orders:
    summary: CRUD Operations for Order History
    get:
      summary: Get all order history
      description: ''
      operationId: get_orders
      responses:
        '200':
          description: A User object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The order ID.
                  user_id:
                    type: integer
                    description: Id representing the owner of this order.
                  cart_id:
                    type: integer
                    description: Id representing the owner's cart which makes up the order.
                  price_total:
                    type: integer
                    description: Amount of all the items in the cart.
                  complete:
                    type: boolean
                    description: If the order has been completed or not.
                  completed_at:
                    type: integer
                    description: Timestamp of when the order was completed if it was.
        default:
          description: Default error sample response
      tags:
        - Orders
    post:
      summary: 'Add a new order '
      description: ''
      requestBody:
          description: Updating an item in a user's cart
          content:
            application/json: 
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    description: Id representing the owner of this order.
                  cart_id:
                    type: integer
                    description: Id representing the owner's cart which makes up the order.
                  price_total:
                    type: integer
                    description: Amount of all the items in the cart.
                  complete:
                    type: boolean
                    description: If the order has been completed or not.
                  completed_at:
                    type: integer
                    description: Timestamp of when the order was completed if it was.
      operationId: add_order
      responses:
        '200':
          description: A User object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The order ID.
                  user_id:
                    type: integer
                    description: Id representing the owner of this order.
                  cart_id:
                    type: integer
                    description: Id representing the owner's cart which makes up the order.
                  price_total:
                    type: integer
                    description: Amount of all the items in the cart.
                  complete:
                    type: boolean
                    description: If the order has been completed or not.
                  completed_at:
                    type: integer
                    description: Timestamp of when the order was completed if it was.
        default:
          description: Default error sample response
      tags:
        - Orders
  /api/orders/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: {}
    get:
      summary: Get a specific order
      description: ''
      operationId: get_specific_order
      responses:
        '200':
          description: A User object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The order ID.
                  user_id:
                    type: integer
                    description: Id representing the owner of this order.
                  cart_id:
                    type: integer
                    description: Id representing the owner's cart which makes up the order.
                  price_total:
                    type: integer
                    description: Amount of all the items in the cart.
                  complete:
                    type: boolean
                    description: If the order has been completed or not.
                  completed_at:
                    type: integer
                    description: Timestamp of when the order was completed if it was.
        default:
          description: Default error sample response
      tags:
        - Orders
    put:
      summary: Update existing order
      description: ''
      operationId: Update_order
      requestBody:
          description: Updating an item in a user's cart
          content:
            application/json: 
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                    description: Id representing the owner of this order.
                  cart_id:
                    type: integer
                    description: Id representing the owner's cart which makes up the order.
                  price_total:
                    type: integer
                    description: Amount of all the items in the cart.
                  complete:
                    type: boolean
                    description: If the order has been completed or not.
                  completed_at:
                    type: integer
                    description: Timestamp of when the order was completed if it was.
      responses:
        '200':
          description: A User object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The order ID.
                  user_id:
                    type: integer
                    description: Id representing the owner of this order.
                  cart_id:
                    type: integer
                    description: Id representing the owner's cart which makes up the order.
                  price_total:
                    type: integer
                    description: Amount of all the items in the cart.
                  complete:
                    type: boolean
                    description: If the order has been completed or not.
                  completed_at:
                    type: integer
                    description: Timestamp of when the order was completed if it was.
        default:
          description: Default error sample response
      tags:
        - Orders
    delete:
      summary: Delete specific order
      description: ''
      operationId: delete_order
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Success:
                  value: Order Successfully Deleted
        default:
          description: Default error sample response
      tags:
        - Orders